refactor: split services/spotify.ts into focused modules
Split the monolithic 1,197-line services/spotify.ts into the following focused modules:
- src/services/spotify/types.ts: Type definitions (TokenData, Track, PlaylistInfo, AlbumInfo, SpotifyImage, etc.)
- src/services/spotify/api.ts: Fetch wrapper with rate limiting and request deduplication
- src/services/spotify/cache.ts: In-memory cache state and invalidation functions
- src/services/spotify/auth.ts: SpotifyAuth class for OAuth 2.0 PKCE authentication
- src/services/spotify/tracks.ts: Track-related functions (getLikedSongs, checkTrackSaved, saveTrack, etc.)
- src/services/spotify/playlists.ts: Playlist functions (getPlaylistTracks, getUserLibraryInterleaved, etc.)
- src/services/spotify/albums.ts: Album functions (getAlbumTracks, checkAlbumSaved, saveAlbum, etc.)
- src/services/spotify/search.ts: Track search functionality
- src/services/spotify/playlists-crud.ts: Playlist creation and management (createPlaylist, addTracksToPlaylist, etc.)
- src/services/spotify/index.ts: Barrel export re-exporting the full public API

The original src/services/spotify.ts is replaced with a forwarding shim to maintain backward compatibility.
This refactoring maintains all existing functionality while organizing code by responsibility.

Resolves: #392
