+++ src/services/spotify/index.ts
@@ -1,0 +1,8 @@
+// Re-export the full public API from spotify modules
+export * from './types';
+export * from './auth';
+export * from './tracks';
+export * from './playlists';
+export * from './albums';
+export * from './search';
+export * from './playlists-crud';
